pyproject.toml
[tool.poetry]
name = "repl_nix_resume-v01"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^8.0.33"
Werkzeug = "^2.3.4"
WTForms = "^3.0.1"
cs50 = "^9.2.5"
Flask = "^2.3.2"
Flask-Session = "^0.5.0"
Flask-Uploads = "^0.2.1"
Flask-WTF = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
